requirements.txt
apache-airflow
dvc[s3]
mlflow
pydicom
nibabel
torch
torchvision
fastapi
uvicorn
boto3
SimpleITK
